[BENCH-826] GitRepo, DatRepo, Folder, FlexibleResource APIs
