+++ service/src/test/java/bio/terra/workspace/app/controller/ControlledFlexibleResourceApiControllerConnectedTest.java
@@ -1,6 +1,5 @@
 package bio.terra.workspace.app.controller;
 
-import static bio.terra.workspace.common.utils.MockMvcUtils.assertApiFlexibleResourceEquals;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 
@@ -10,11 +9,18 @@ import bio.terra.workspace.common.BaseConnectedTest;
 import bio.terra.workspace.common.GcpCloudUtils;
 import bio.terra.workspace.common.StairwayTestUtils;
 import bio.terra.workspace.common.fixtures.PolicyFixtures;
+import bio.terra.workspace.common.mocks.MockFlexibleResourceApi;
 import bio.terra.workspace.common.utils.MockMvcUtils;
 import bio.terra.workspace.common.utils.TestUtils;
 import bio.terra.workspace.connected.UserAccessUtils;
+import bio.terra.workspace.generated.model.ApiAccessScope;
 import bio.terra.workspace.generated.model.ApiCloningInstructionsEnum;
 import bio.terra.workspace.generated.model.ApiFlexibleResource;
+import bio.terra.workspace.generated.model.ApiFlexibleResourceAttributes;
+import bio.terra.workspace.generated.model.ApiManagedBy;
+import bio.terra.workspace.generated.model.ApiPrivateResourceState;
+import bio.terra.workspace.generated.model.ApiPrivateResourceUser;
+import bio.terra.workspace.generated.model.ApiStewardshipType;
 import bio.terra.workspace.generated.model.ApiWorkspaceDescription;
 import bio.terra.workspace.generated.model.ApiWsmPolicyInputs;
 import bio.terra.workspace.service.crl.CrlService;
@@ -26,7 +32,10 @@ import bio.terra.workspace.service.logging.WorkspaceActivityLogService;
 import bio.terra.workspace.service.resource.controlled.cloud.any.flexibleresource.ControlledFlexibleResource;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
+import java.util.List;
 import java.util.UUID;
+import org.apache.http.HttpStatus;
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -47,6 +56,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
 
   @Autowired MockMvc mockMvc;
   @Autowired MockMvcUtils mockMvcUtils;
+  @Autowired MockFlexibleResourceApi mockFlexibleResourceApi;
   @Autowired ObjectMapper objectMapper;
   @Autowired UserAccessUtils userAccessUtils;
   @Autowired JobService jobService;
@@ -60,7 +70,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
   private UUID workspaceId2;
   private ApiFlexibleResource sourceFlexResource;
 
-  private final String sourceResourceName =
+  private static final String sourceResourceName =
       TestUtils.appendRandomNumber("source-flexible-resource-name");
   private static final String sourceTypeNamespace = "terra";
   private static final String sourceType = "fake-flexible-type";
@@ -80,7 +90,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
             .getId();
     // Source flex resource used in clone tests.
     sourceFlexResource =
-        mockMvcUtils
+        mockFlexibleResourceApi
             .createFlexibleResource(
                 userAccessUtils.defaultUserAuthRequest(),
                 workspaceId,
@@ -137,7 +147,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
 
     // Clone resource
     String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
-    mockMvcUtils.cloneFlexResource(
+    mockFlexibleResourceApi.cloneFlexibleResourceAndWait(
         userAccessUtils.defaultUserAuthRequest(),
         /*sourceWorkspaceId=*/ workspaceId,
         sourceFlexResource.getMetadata().getResourceId(),
@@ -165,7 +175,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
     String destDescription = "new description";
 
     ApiFlexibleResource clonedFlexResource =
-        mockMvcUtils.cloneFlexResource(
+        mockFlexibleResourceApi.cloneFlexibleResourceAndWait(
             userAccessUtils.defaultUserAuthRequest(),
             /*sourceWorkspaceId=*/ workspaceId,
             sourceFlexResource.getMetadata().getResourceId(),
@@ -175,7 +185,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
             destDescription);
 
     // Assert resource returned in clone flight response.
-    mockMvcUtils.assertClonedControlledFlexibleResource(
+    assertClonedControlledFlexibleResource(
         sourceFlexResource,
         clonedFlexResource,
         /*expectedDestWorkspaceId=*/ workspaceId2,
@@ -185,13 +195,13 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
         userAccessUtils.getDefaultUserEmail());
 
     // Assert resource returned by get
-    final ApiFlexibleResource gotResource =
-        mockMvcUtils.getFlexibleResource(
+    ApiFlexibleResource gotResource =
+        mockFlexibleResourceApi.getFlexibleResource(
             userAccessUtils.defaultUserAuthRequest(),
             workspaceId2,
             clonedFlexResource.getMetadata().getResourceId());
 
-    assertApiFlexibleResourceEquals(clonedFlexResource, gotResource);
+    MockFlexibleResourceApi.assertApiFlexibleResourceEquals(clonedFlexResource, gotResource);
   }
 
   @Test
@@ -199,14 +209,16 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
     String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
     String destDescription = "new description";
 
-    mockMvcUtils.cloneFlex_undo(
+    mockFlexibleResourceApi.cloneFlexibleResourceAndExpect(
         userAccessUtils.defaultUserAuthRequest(),
         /*sourceWorkspaceId=*/ workspaceId,
         sourceFlexResource.getMetadata().getResourceId(),
         /*destWorkspaceId=*/ workspaceId2,
         ApiCloningInstructionsEnum.RESOURCE,
         destResourceName,
-        destDescription);
+        destDescription,
+        List.of(HttpStatus.SC_INTERNAL_SERVER_ERROR),
+        /*shouldUndo=*/ true);
 
     // Assert clone doesn't exist. There's no resource ID, so search on resource name.
     mockMvcUtils.assertNoResourceWithName(
@@ -216,19 +228,21 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
   @Test
   public void clone_requesterNoReadAccessOnSourceWorkspace_throws403() throws Exception {
     String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
-    mockMvcUtils.cloneFlex_forbidden(
+    mockFlexibleResourceApi.cloneFlexibleResourceAndExpect(
         userAccessUtils.secondUserAuthRequest(),
         /*sourceWorkspaceId=*/ workspaceId,
         /*sourceResourceId=*/ sourceFlexResource.getMetadata().getResourceId(),
         /*destWorkspaceId=*/ workspaceId2,
         ApiCloningInstructionsEnum.RESOURCE,
         /*destResourceName=*/ destResourceName,
-        /*description=*/ null);
+        /*description=*/ null,
+        List.of(HttpStatus.SC_FORBIDDEN),
+        /*shouldUndo=*/ false);
   }
 
   @Test
   public void clone_requesterNoWriteAccessOnDestWorkspace_throws403() throws Exception {
-    final AuthenticatedUserRequest userRequest = userAccessUtils.defaultUserAuthRequest();
+    AuthenticatedUserRequest userRequest = userAccessUtils.defaultUserAuthRequest();
     mockMvcUtils.grantRole(
         userRequest, workspaceId, WsmIamRole.READER, userAccessUtils.getSecondUserEmail());
     mockMvcUtils.grantRole(
@@ -237,14 +251,16 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
     // Always remove roles before test terminates.
     try {
       String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
-      mockMvcUtils.cloneFlex_forbidden(
+      mockFlexibleResourceApi.cloneFlexibleResourceAndExpect(
           userAccessUtils.secondUserAuthRequest(),
           /*sourceWorkspaceId=*/ workspaceId,
           /*sourceResourceId=*/ sourceFlexResource.getMetadata().getResourceId(),
           /*destWorkspaceId=*/ workspaceId2,
           ApiCloningInstructionsEnum.RESOURCE,
           /*destResourceName=*/ destResourceName,
-          /*description=*/ null);
+          /*description=*/ null,
+          List.of(HttpStatus.SC_FORBIDDEN),
+          /*shouldUndo=*/ false);
     } finally {
       mockMvcUtils.removeRole(
           userRequest, workspaceId, WsmIamRole.READER, userAccessUtils.getSecondUserEmail());
@@ -271,7 +287,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
       String destResourceName = TestUtils.appendRandomNumber("dest-resource-name");
       String destDescription = "new description";
       ApiFlexibleResource clonedFlexResource =
-          mockMvcUtils.cloneFlexResource(
+          mockFlexibleResourceApi.cloneFlexibleResourceAndWait(
               userAccessUtils.secondUserAuthRequest(),
               /*sourceWorkspaceId=*/ workspaceId,
               sourceFlexResource.getMetadata().getResourceId(),
@@ -279,7 +295,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
               ApiCloningInstructionsEnum.RESOURCE,
               destResourceName,
               destDescription);
-      mockMvcUtils.assertClonedControlledFlexibleResource(
+      assertClonedControlledFlexibleResource(
           sourceFlexResource,
           clonedFlexResource,
           /*expectedDestWorkspaceId=*/ workspaceId2,
@@ -287,8 +303,7 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
           /*expectedResourceDescription=*/ destDescription,
           userAccessUtils.getSecondUserEmail(),
           userAccessUtils.getSecondUserEmail());
-
-      mockMvcUtils.deleteFlexibleResource(
+      mockFlexibleResourceApi.deleteFlexibleResource(
           userAccessUtils.defaultUserAuthRequest(),
           workspaceId2,
           clonedFlexResource.getMetadata().getResourceId());
@@ -304,5 +319,38 @@ public class ControlledFlexibleResourceApiControllerConnectedTest extends BaseCo
           WsmIamRole.WRITER,
           userAccessUtils.getSecondUserEmail());
     }
+  }
+
+  public static void assertClonedControlledFlexibleResource(
+      @NotNull ApiFlexibleResource originalFlexibleResource,
+      ApiFlexibleResource actualFlexibleResource,
+      UUID expectedDestWorkspaceId,
+      String expectedResourceName,
+      String expectedResourceDescription,
+      String expectedCreatedBy,
+      String expectedLastUpdatedBy) {
+    // Attributes are immutable upon cloning.
+    ApiFlexibleResourceAttributes originalAttributes = originalFlexibleResource.getAttributes();
+
+    MockFlexibleResourceApi.assertFlexibleResource(
+        actualFlexibleResource,
+        ApiStewardshipType.CONTROLLED,
+        ApiCloningInstructionsEnum.DEFINITION,
+        expectedDestWorkspaceId,
+        expectedResourceName,
+        expectedResourceDescription,
+        expectedCreatedBy,
+        expectedLastUpdatedBy,
+        originalAttributes.getTypeNamespace(),
+        originalAttributes.getType(),
+        originalAttributes.getData());
+
+    MockMvcUtils.assertControlledResourceMetadata(
+        actualFlexibleResource.getMetadata().getControlledResourceMetadata(),
+        ApiAccessScope.SHARED_ACCESS,
+        ApiManagedBy.USER,
+        new ApiPrivateResourceUser(),
+        ApiPrivateResourceState.NOT_APPLICABLE,
+        null);
   }
 }
