+++ service/src/test/java/bio/terra/workspace/service/resource/statetests/AnyResourceStateFailureTest.java
@@ -1,16 +1,16 @@
 package bio.terra.workspace.service.resource.statetests;
 
+import static bio.terra.workspace.common.mocks.MockDataRepoApi.CREATE_REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT;
+import static bio.terra.workspace.common.mocks.MockDataRepoApi.REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT;
+import static bio.terra.workspace.common.mocks.MockFlexibleResourceApi.CREATE_CONTROLLED_FLEXIBLE_RESOURCES_PATH_FORMAT;
+import static bio.terra.workspace.common.mocks.MockGitRepoApi.CREATE_REFERENCED_GIT_REPOS_PATH_FORMAT;
+import static bio.terra.workspace.common.mocks.MockGitRepoApi.REFERENCED_GIT_REPOS_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.CREATE_REFERENCED_GCP_BQ_DATASETS_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.CREATE_REFERENCED_GCP_BQ_DATA_TABLES_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.CREATE_REFERENCED_GCP_GCS_BUCKETS_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.CREATE_REFERENCED_GCP_GCS_OBJECTS_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.REFERENCED_GCP_BQ_DATASET_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockGcpApi.REFERENCED_GCP_BQ_DATA_TABLE_PATH_FORMAT;
-import static bio.terra.workspace.common.utils.MockMvcUtils.CONTROLLED_FLEXIBLE_RESOURCES_V1_PATH_FORMAT;
-import static bio.terra.workspace.common.utils.MockMvcUtils.REFERENCED_DATA_REPO_SNAPSHOTS_V1_PATH_FORMAT;
-import static bio.terra.workspace.common.utils.MockMvcUtils.REFERENCED_DATA_REPO_SNAPSHOT_V1_PATH_FORMAT;
-import static bio.terra.workspace.common.utils.MockMvcUtils.REFERENCED_GIT_REPOS_V1_PATH_FORMAT;
-import static bio.terra.workspace.common.utils.MockMvcUtils.REFERENCED_GIT_REPO_V1_PATH_FORMAT;
 import static bio.terra.workspace.common.utils.MockMvcUtils.USER_REQUEST;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -19,9 +19,11 @@ import bio.terra.workspace.common.BaseUnitTest;
 import bio.terra.workspace.common.fixtures.ControlledResourceFixtures;
 import bio.terra.workspace.common.fixtures.ReferenceResourceFixtures;
 import bio.terra.workspace.common.fixtures.WorkspaceFixtures;
+import bio.terra.workspace.common.mocks.MockFlexibleResourceApi;
 import bio.terra.workspace.common.utils.MockMvcUtils;
 import bio.terra.workspace.db.ResourceDao;
 import bio.terra.workspace.db.WorkspaceDao;
+import bio.terra.workspace.generated.model.ApiCreateControlledFlexibleResourceRequestBody;
 import bio.terra.workspace.generated.model.ApiCreateDataRepoSnapshotReferenceRequestBody;
 import bio.terra.workspace.generated.model.ApiCreateGcpBigQueryDataTableReferenceRequestBody;
 import bio.terra.workspace.generated.model.ApiCreateGcpBigQueryDatasetReferenceRequestBody;
@@ -55,6 +57,7 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
 
   @Autowired private MockMvc mockMvc;
   @Autowired private MockMvcUtils mockMvcUtils;
+  @Autowired private MockFlexibleResourceApi mockFlexibleResourceApi;
   @Autowired ObjectMapper objectMapper;
   @Autowired ReferencedResourceService referencedResourceService;
   @Autowired ResourceDao resourceDao;
@@ -88,13 +91,13 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
     workspaceDao.createWorkspaceStart(workspace, /* applicationIds */ null, flightId);
 
     // ANY-Controlled Flexible
-    var flexRequest =
-        mockMvcUtils.createFlexibleResourceRequestBody(
+    ApiCreateControlledFlexibleResourceRequestBody flexRequest =
+        mockFlexibleResourceApi.createFlexibleResourceRequestBody(
             RESOURCE_NAME, "terra", "footype", "foodata".getBytes(StandardCharsets.UTF_8));
     mockMvcUtils.postExpect(
         USER_REQUEST,
         objectMapper.writeValueAsString(flexRequest),
-        CONTROLLED_FLEXIBLE_RESOURCES_V1_PATH_FORMAT.formatted(workspace.workspaceId()),
+        CREATE_CONTROLLED_FLEXIBLE_RESOURCES_PATH_FORMAT.formatted(workspace.workspaceId()),
         HttpStatus.SC_CONFLICT);
 
     // ANY-Referenced Data Repo Snapshot
@@ -110,7 +113,7 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
     mockMvcUtils.postExpect(
         USER_REQUEST,
         objectMapper.writeValueAsString(tdrRequest),
-        REFERENCED_DATA_REPO_SNAPSHOTS_V1_PATH_FORMAT.formatted(workspace.workspaceId()),
+        CREATE_REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT.formatted(workspace.workspaceId()),
         HttpStatus.SC_CONFLICT);
 
     // ANY-Referenced Git Repo
@@ -123,7 +126,7 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
     mockMvcUtils.postExpect(
         USER_REQUEST,
         objectMapper.writeValueAsString(gitRequest),
-        REFERENCED_GIT_REPOS_V1_PATH_FORMAT.formatted(workspace.workspaceId()),
+        CREATE_REFERENCED_GIT_REPOS_PATH_FORMAT.formatted(workspace.workspaceId()),
         HttpStatus.SC_CONFLICT);
 
     // GCP-Referenced BQ dataset
@@ -228,7 +231,7 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
     workspaceDao.deleteWorkspaceStart(workspaceUuid, flightId);
 
     // ANY-Controlled Flexible
-    mockMvcUtils.updateFlexibleResourceExpect(
+    mockFlexibleResourceApi.updateFlexibleResourceExpect(
         workspaceUuid,
         flexResource.getResourceId(),
         null,
@@ -237,27 +240,27 @@ public class AnyResourceStateFailureTest extends BaseUnitTest {
         null,
         HttpStatus.SC_CONFLICT);
     stateTestUtils.deleteResourceExpectConflict(
-        workspaceUuid, flexResource.getResourceId(), REFERENCED_DATA_REPO_SNAPSHOT_V1_PATH_FORMAT);
+        workspaceUuid, flexResource.getResourceId(), REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT);
 
     // ANY-Referenced Data Repo Snapshot
     var tdrUpdateRequest = new ApiUpdateDataRepoSnapshotReferenceRequestBody().snapshot("foobar");
     stateTestUtils.postResourceExpectConflict(
         workspaceUuid,
         tdrResource.getResourceId(),
-        REFERENCED_DATA_REPO_SNAPSHOT_V1_PATH_FORMAT,
+        REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT,
         objectMapper.writeValueAsString(tdrUpdateRequest));
     stateTestUtils.deleteResourceExpectConflict(
-        workspaceUuid, tdrResource.getResourceId(), REFERENCED_DATA_REPO_SNAPSHOT_V1_PATH_FORMAT);
+        workspaceUuid, tdrResource.getResourceId(), REFERENCED_DATA_REPO_SNAPSHOTS_PATH_FORMAT);
 
     // ANY-Referenced Git Repo
     var gitUpdateRequest = new ApiUpdateGitRepoReferenceRequestBody().gitRepoUrl("new-fake-url");
     stateTestUtils.patchResourceExpectConflict(
         workspaceUuid,
         gitResource.getResourceId(),
-        REFERENCED_GIT_REPO_V1_PATH_FORMAT,
+        REFERENCED_GIT_REPOS_PATH_FORMAT,
         objectMapper.writeValueAsString(gitUpdateRequest));
     stateTestUtils.deleteResourceExpectConflict(
-        workspaceUuid, gitResource.getResourceId(), REFERENCED_GIT_REPO_V1_PATH_FORMAT);
+        workspaceUuid, gitResource.getResourceId(), REFERENCED_GIT_REPOS_PATH_FORMAT);
 
     // GCP-Referenced BQ dataset
     var bqUpdateRequest = new ApiUpdateBigQueryDatasetReferenceRequestBody().datasetId("foo");
